app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sltptr <mike@sltptr.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
sqlalchemy = "^2.0.31"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
transformers = {extras = ["torch"], version = "^4.42.4"}
torch = "^2.3.1"
feedparser = "^6.0.11"
pydantic = "^2.8.2"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
sagemaker = ">=2.140.0"
datasets = {version = ">=2.10.0", extras = ["s3fs"]}
s3fs = "^2024.6.1"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
arrow = "^1.3.0"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.2"

[tool.poetry.scripts]
start = "app.entrypoint:main"
generate = "app.jobs.generate:main"
tfidf = "app.jobs.tfidf:main"
distilbert = "app.jobs.distilbert.entrypoint:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
